Move towards storing and archiving refresh batches on disk
Instead of storing the refresh batches in a relational database, where
useless indexes and relation tracking bloats the database to tens of
Gigabytes, we can store the batches as-is in the JSON format. This JSON
has the same schema as the JSON provided by doomlister. For the most
part, the records in the database aren't used directly for presentation.
Instead, they are used to generate the presentation tables. Therefore,
we can just as well store them in unbloated, compressed files.

The 'db_to_json' command will extract all batches from the database
and dump it in the target JSON. This is followed up by 'archive_batches'
command that takes the dumped JSON and puts it into monthly tarballs.
Batches from the current month will never be tarballed as they are still
being collected and we don't want an incomplete archive.