Thin the database; remove the batch data that is not being read
All of the server information that is not being used to display
the pages is removed from the database. This data from now on should
be stored only as archived JSON files. If any new presentation needs
to be added, these JSON files can be used to create that presentation.